Feat: 댓글 신고, 선호 도시 조회/추가/삭제 API 구현

+++ src/main/java/com/jandi/plan_backend/commu/controller/CommentController.java
@@ -1,12 +1,10 @@
 package com.jandi.plan_backend.commu.controller;
 
-import com.jandi.plan_backend.commu.dto.CommentReqDTO;
-import com.jandi.plan_backend.commu.dto.CommentRespDTO;
-import com.jandi.plan_backend.commu.dto.ParentCommentDTO;
-import com.jandi.plan_backend.commu.dto.RepliesDTO;
+import com.jandi.plan_backend.commu.dto.*;
 import com.jandi.plan_backend.commu.service.CommentService;
 import com.jandi.plan_backend.security.JwtTokenProvider;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -70,10 +68,12 @@ public class CommentController {
             @RequestHeader("Authorization") String token, // 헤더의 Authorization에서 JWT 토큰 받기
             @RequestBody CommentReqDTO commentDTO // JSON 형식으로 게시글 작성 정보 받기
     ){
-
         // Jwt 토큰으로부터 유저 이메일 추출
         String jwtToken = token.replace("Bearer ", "");
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
+        if (userEmail == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
         // 댓글 저장 및 반환
         CommentRespDTO savedComment = commentService.writeComment(commentDTO, postId, userEmail);
@@ -87,10 +87,12 @@ public class CommentController {
             @RequestHeader("Authorization") String token, // 헤더의 Authorization에서 JWT 토큰 받기
             @RequestBody CommentReqDTO commentDTO // JSON 형식으로 게시글 작성 정보 받기
     ){
-
         // Jwt 토큰으로부터 유저 이메일 추출
         String jwtToken = token.replace("Bearer ", "");
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
+        if (userEmail == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
         // 댓글 저장 및 반환
         CommentRespDTO savedComment = commentService.writeReplies(commentDTO, commentId, userEmail);
@@ -107,6 +109,9 @@ public class CommentController {
         // Jwt 토큰으로부터 유저 이메일 추출
         String jwtToken = token.replace("Bearer ", "");
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
+        if (userEmail == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
         // 게시물 수정 및 반환
         CommentRespDTO updatedPost = commentService.updateComment(commentDTO, commentId, userEmail);
@@ -122,6 +127,9 @@ public class CommentController {
         // Jwt 토큰으로부터 유저 이메일 추출
         String jwtToken = token.replace("Bearer ", "");
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
+        if (userEmail == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
         // 답글 삭제 및 반환
         int deletedRepliesCount = (commentService.deleteComments(userEmail, commentId));
@@ -158,4 +166,20 @@ public class CommentController {
         commentService.deleteLikeComment(userEmail, commentId);
         return ResponseEntity.ok("좋아요 취소되었습니다.");
     }
+
+    /** 댓글 신고 API */
+    @PostMapping("/comments/reports/{commentId}")
+    public ResponseEntity<?> reportComment(
+            @PathVariable Integer commentId,
+            @RequestBody ReportReqDTO reportDTO,
+            @RequestHeader("Authorization") String token // 헤더의 Authorization에서 JWT 토큰 받기
+    ){
+        // Jwt 토큰으로부터 유저 이메일 추출
+        String jwtToken = token.replace("Bearer ", "");
+        String userEmail = jwtTokenProvider.getEmail(jwtToken);
+
+        CommentReportRespDTO reported = commentService.reportComment(userEmail, commentId, reportDTO);
+        return ResponseEntity.ok(reported);
+    }
+
 }
